# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SmartHire API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: interviews
    description: Interview management
  - name: scorecards
    description: Scorecard retrieval
  - name: bot
    description: Teams bot integration
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /interviews:
    get:
      operationId: listInterviews
      tags: [interviews]
      summary: List all interviews
      parameters:
        - name: source
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of interviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Interview"
    post:
      operationId: createInterview
      tags: [interviews]
      summary: Create interview and generate questions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInterviewBody"
      responses:
        "201":
          description: Interview created with generated questions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewWithQuestions"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /interviews/stats:
    get:
      operationId: getInterviewStats
      tags: [interviews]
      summary: Get interview statistics
      responses:
        "200":
          description: Interview statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewStats"

  /interviews/{id}:
    get:
      operationId: getInterview
      tags: [interviews]
      summary: Get a single interview
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Interview details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewWithQuestions"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /interviews/{id}/submit:
    post:
      operationId: submitInterview
      tags: [interviews]
      summary: Submit candidate answers and trigger evaluation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitInterviewBody"
      responses:
        "200":
          description: Evaluation complete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scorecard"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /interviews/{id}/scorecard:
    get:
      operationId: getScorecard
      tags: [scorecards]
      summary: Get scorecard for an interview
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Scorecard data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScorecardWithDetails"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /bot/health:
    get:
      operationId: getBotHealth
      tags: [bot]
      summary: Bot health check
      responses:
        "200":
          description: Bot health status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotHealthStatus"

  /bot/submit-interview:
    post:
      operationId: botSubmitInterview
      tags: [bot]
      summary: Teams bot submit interview
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BotSubmitInterviewBody"
      responses:
        "200":
          description: Interview processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotSubmitInterviewResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Interview:
      type: object
      properties:
        id:
          type: integer
        recruiterName:
          type: string
        candidateName:
          type: string
        candidateEmail:
          type: string
        jobTitle:
          type: string
        jobDescription:
          type: string
        status:
          type: string
          enum: [pending, in_progress, completed, evaluating]
        overallScore:
          type: ["number", "null"]
        verdict:
          type: ["string", "null"]
        createdAt:
          type: string
        completedAt:
          type: ["string", "null"]
        duration:
          type: ["integer", "null"]
        llmUsed:
          type: ["string", "null"]
        source:
          type: ["string", "null"]
      required:
        - id
        - recruiterName
        - candidateName
        - candidateEmail
        - jobTitle
        - jobDescription
        - status
        - overallScore
        - verdict
        - createdAt
        - completedAt
        - duration
        - llmUsed
        - source

    Question:
      type: object
      properties:
        id:
          type: integer
        interviewId:
          type: integer
        questionIndex:
          type: integer
        questionText:
          type: string
        questionType:
          type: string
          enum: [technical, behavioral, situational]
      required:
        - id
        - interviewId
        - questionIndex
        - questionText
        - questionType

    Answer:
      type: object
      properties:
        id:
          type: integer
        questionId:
          type: integer
        questionIndex:
          type: integer
        answerText:
          type: string
        score:
          type: ["number", "null"]
        feedback:
          type: ["string", "null"]
        confidenceScore:
          type: ["number", "null"]
        fillerWordCount:
          type: ["integer", "null"]
        pauseCount:
          type: ["integer", "null"]
        speechDurationSeconds:
          type: ["integer", "null"]
      required:
        - id
        - questionId
        - questionIndex
        - answerText
        - score
        - feedback
        - confidenceScore
        - fillerWordCount
        - pauseCount
        - speechDurationSeconds

    InterviewWithQuestions:
      type: object
      properties:
        id:
          type: integer
        recruiterName:
          type: string
        candidateName:
          type: string
        candidateEmail:
          type: string
        jobTitle:
          type: string
        jobDescription:
          type: string
        status:
          type: string
          enum: [pending, in_progress, completed, evaluating]
        overallScore:
          type: ["number", "null"]
        verdict:
          type: ["string", "null"]
        createdAt:
          type: string
        completedAt:
          type: ["string", "null"]
        duration:
          type: ["integer", "null"]
        llmUsed:
          type: ["string", "null"]
        source:
          type: ["string", "null"]
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
      required:
        - id
        - recruiterName
        - candidateName
        - candidateEmail
        - jobTitle
        - jobDescription
        - status
        - overallScore
        - verdict
        - createdAt
        - completedAt
        - duration
        - llmUsed
        - source
        - questions

    Scorecard:
      type: object
      properties:
        id:
          type: integer
        interviewId:
          type: integer
        technicalScore:
          type: number
        communicationScore:
          type: number
        problemSolvingScore:
          type: number
        roleRelevanceScore:
          type: number
        speechConfidenceScore:
          type: ["number", "null"]
        overallScore:
          type: number
        verdict:
          type: string
        strengths:
          type: array
          items:
            type: string
        improvements:
          type: array
          items:
            type: string
        summary:
          type: string
        recruiterNote:
          type: string
        createdAt:
          type: string
      required:
        - id
        - interviewId
        - technicalScore
        - communicationScore
        - problemSolvingScore
        - roleRelevanceScore
        - speechConfidenceScore
        - overallScore
        - verdict
        - strengths
        - improvements
        - summary
        - recruiterNote
        - createdAt

    ScorecardWithDetails:
      type: object
      properties:
        scorecard:
          $ref: "#/components/schemas/Scorecard"
        interview:
          $ref: "#/components/schemas/Interview"
        answers:
          type: array
          items:
            $ref: "#/components/schemas/Answer"
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
      required:
        - scorecard
        - interview
        - answers
        - questions

    AnswerInput:
      type: object
      properties:
        questionIndex:
          type: integer
        answerText:
          type: string
        confidenceScore:
          type: ["number", "null"]
        fillerWordCount:
          type: ["integer", "null"]
        pauseCount:
          type: ["integer", "null"]
        speechDurationSeconds:
          type: ["integer", "null"]
      required:
        - questionIndex
        - answerText

    CreateInterviewBody:
      type: object
      properties:
        recruiterName:
          type: string
        candidateName:
          type: string
        candidateEmail:
          type: string
        jobTitle:
          type: string
        jobDescription:
          type: string
      required:
        - recruiterName
        - candidateName
        - candidateEmail
        - jobTitle
        - jobDescription

    SubmitInterviewBody:
      type: object
      properties:
        answers:
          type: array
          items:
            $ref: "#/components/schemas/AnswerInput"
      required:
        - answers

    InterviewStats:
      type: object
      properties:
        total:
          type: integer
        completed:
          type: integer
        pending:
          type: integer
        averageScore:
          type: ["number", "null"]
        verdictBreakdown:
          type: object
          properties:
            strongHire:
              type: integer
            hire:
              type: integer
            maybe:
              type: integer
            noHire:
              type: integer
          required:
            - strongHire
            - hire
            - maybe
            - noHire
      required:
        - total
        - completed
        - pending
        - averageScore
        - verdictBreakdown

    BotHealthStatus:
      type: object
      properties:
        status:
          type: string
        ollamaAvailable:
          type: boolean
        gptAvailable:
          type: boolean
        timestamp:
          type: string
      required:
        - status
        - ollamaAvailable
        - gptAvailable
        - timestamp

    BotAnswerInput:
      type: object
      properties:
        questionText:
          type: string
        answerText:
          type: string
        confidenceScore:
          type: ["number", "null"]
        fillerWordCount:
          type: ["integer", "null"]
        pauseCount:
          type: ["integer", "null"]
        speechDurationSeconds:
          type: ["integer", "null"]
      required:
        - questionText
        - answerText

    BotSubmitInterviewBody:
      type: object
      properties:
        candidateName:
          type: string
        recruiterEmail:
          type: string
        jobTitle:
          type: string
        jobDescription:
          type: string
        answers:
          type: array
          items:
            $ref: "#/components/schemas/BotAnswerInput"
      required:
        - candidateName
        - recruiterEmail
        - jobTitle
        - jobDescription
        - answers

    BotSubmitInterviewResponse:
      type: object
      properties:
        interviewId:
          type: integer
        scorecardId:
          type: integer
        scorecardUrl:
          type: string
      required:
        - interviewId
        - scorecardId
        - scorecardUrl