# lib/api-zod/src/generated/types/answer.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SmartHire API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Answer {
  id: number;
  questionId: number;
  questionIndex: number;
  answerText: string;
  /** @nullable */
  score: number | null;
  /** @nullable */
  feedback: string | null;
  /** @nullable */
  confidenceScore: number | null;
  /** @nullable */
  fillerWordCount: number | null;
  /** @nullable */
  pauseCount: number | null;
  /** @nullable */
  speechDurationSeconds: number | null;
}