# lib/api-zod/src/generated/types/botAnswerInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SmartHire API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BotAnswerInput {
  questionText: string;
  answerText: string;
  /** @nullable */
  confidenceScore?: number | null;
  /** @nullable */
  fillerWordCount?: number | null;
  /** @nullable */
  pauseCount?: number | null;
  /** @nullable */
  speechDurationSeconds?: number | null;
}