# lib/api-zod/src/generated/types/botHealthStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SmartHire API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BotHealthStatus {
  status: string;
  ollamaAvailable: boolean;
  gptAvailable: boolean;
  timestamp: string;
}