# lib/api-zod/src/generated/types/botSubmitInterviewBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SmartHire API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BotAnswerInput } from "./botAnswerInput";

export interface BotSubmitInterviewBody {
  candidateName: string;
  recruiterEmail: string;
  jobTitle: string;
  jobDescription: string;
  answers: BotAnswerInput[];
}