# lib/api-zod/src/generated/types/botSubmitInterviewResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SmartHire API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BotSubmitInterviewResponse {
  interviewId: number;
  scorecardId: number;
  scorecardUrl: string;
}