# lib/api-zod/src/generated/types/interview.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SmartHire API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InterviewStatus } from "./interviewStatus";

export interface Interview {
  id: number;
  recruiterName: string;
  candidateName: string;
  candidateEmail: string;
  jobTitle: string;
  jobDescription: string;
  status: InterviewStatus;
  /** @nullable */
  overallScore: number | null;
  /** @nullable */
  verdict: string | null;
  createdAt: string;
  /** @nullable */
  completedAt: string | null;
  /** @nullable */
  duration: number | null;
  /** @nullable */
  llmUsed: string | null;
  /** @nullable */
  source: string | null;
}