# lib/api-zod/src/generated/types/interviewWithQuestions.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SmartHire API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InterviewWithQuestionsStatus } from "./interviewWithQuestionsStatus";
import type { Question } from "./question";

export interface InterviewWithQuestions {
  id: number;
  recruiterName: string;
  candidateName: string;
  candidateEmail: string;
  jobTitle: string;
  jobDescription: string;
  status: InterviewWithQuestionsStatus;
  /** @nullable */
  overallScore: number | null;
  /** @nullable */
  verdict: string | null;
  createdAt: string;
  /** @nullable */
  completedAt: string | null;
  /** @nullable */
  duration: number | null;
  /** @nullable */
  llmUsed: string | null;
  /** @nullable */
  source: string | null;
  questions: Question[];
}