# lib/api-zod/src/generated/types/scorecard.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SmartHire API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Scorecard {
  id: number;
  interviewId: number;
  technicalScore: number;
  communicationScore: number;
  problemSolvingScore: number;
  roleRelevanceScore: number;
  /** @nullable */
  speechConfidenceScore: number | null;
  overallScore: number;
  verdict: string;
  strengths: string[];
  improvements: string[];
  summary: string;
  recruiterNote: string;
  createdAt: string;
}